Organize eventTemplate into folder
