+++ src/constants/MoveHistory/EventTemplates/index.js
@@ -1,5 +1,5 @@
 export { default as acknowledgeExcessWeightRisk } from './acknowledgeExcessWeightRisk';
-export { default as approveShipment } from './approveShipment';
+export { default as approveShipment } from './ApproveShipment/approveShipment';
 export { default as approveShipmentDiversion } from './approveShipmentDiversion';
 export { default as createBasicServiceItem } from './createBasicServiceItem';
 export { default as createMTOServiceItem } from './createMTOServiceItem';
